test source compile
