+++ core/src/test/java/integration/ElementChecker.java
@@ -1,7 +1,8 @@
 package integration;
 
-import org.thymeleaf.dom.Element;
 import static org.junit.Assert.*;
+
+import org.w3c.dom.Element;
 
 class ElementChecker {
 
@@ -12,7 +13,7 @@ class ElementChecker {
     }
 
     public void containsAttributeWithValue(String attributeName, String attributeValue) {
-        String errorMsg = String.format("Element with id -%s- does not contain attribute -%s- with value -%s-", element.getAttributeValue("id"), attributeName, attributeValue);
-        assertTrue(errorMsg, attributeValue.equals(element.getAttributeValue(attributeName)));
+        String errorMsg = String.format("Element with id -%s- does not contain attribute -%s- with value -%s-", element.getAttribute("id"), attributeName, attributeValue);
+        assertTrue(errorMsg, attributeValue.equals(element.getAttribute(attributeName)));
     }
 }
