+++ core/src/test/java/integration/IntegrationTestBase.java
@@ -1,14 +1,13 @@
 package integration;
 
-import integration.util.ClasspathResourceResolver;
-import java.io.StringReader;
-import net.sourceforge.html5val.Html5ValDialect;
 import org.junit.Before;
 import org.thymeleaf.TemplateEngine;
 import org.thymeleaf.context.Context;
-import org.thymeleaf.dom.Document;
-import org.thymeleaf.templateresolver.TemplateResolver;
-import org.thymeleaf.util.DOMUtils;
+import org.thymeleaf.templateresolver.ClassLoaderTemplateResolver;
+import org.w3c.dom.Document;
+
+import integration.util.XmlUtils;
+import net.sourceforge.html5val.Html5ValDialect;
 
 abstract public class IntegrationTestBase {
 
@@ -16,18 +15,16 @@ abstract public class IntegrationTestBase {
 
     @Before
     public void setUpTemplateEngine() {
-        TemplateResolver templateResolver = new TemplateResolver();
-        templateResolver.setResourceResolver(new ClasspathResourceResolver());
+    	ClassLoaderTemplateResolver templateResolver = new ClassLoaderTemplateResolver();
         templateResolver.setPrefix("/META-INF/");
         templateResolver.setCacheable(false);
         templateEngine = new TemplateEngine();
         templateEngine.setTemplateResolver(templateResolver);
         templateEngine.addDialect(new Html5ValDialect());
-        templateEngine.initialize();
     }
 
     protected Document processTemplate(String templateName, Context context) {
         String html = templateEngine.process(templateName, context);
-        return DOMUtils.getHtml5DOMFor(new StringReader(html));
+        return XmlUtils.toDom(html);
     }
 }
