+++ core/src/test/java/net/sourceforge/html5val/performers/ValidationPerformerFactoryTest.java
@@ -1,14 +1,16 @@
 package net.sourceforge.html5val.performers;
 
+import static org.junit.Assert.*;
+
 import java.lang.annotation.Annotation;
-import net.sourceforge.html5val.performers.ValidationPerformer;
-import net.sourceforge.html5val.performers.ValidationPerformerFactory;
-import net.sourceforge.html5val.performers.EmailPerformer;
+
 import org.hibernate.validator.constraints.Email;
 import org.jmock.Mockery;
 import org.jmock.integration.junit4.JUnit4Mockery;
 import org.junit.Test;
-import static org.junit.Assert.*;
+
+import net.sourceforge.html5val.performers.IValidationPerformer;
+import net.sourceforge.html5val.performers.ValidationPerformerFactory;
 
 // FIXME: only test the "happy path". Test the boundary conditions.
 public class ValidationPerformerFactoryTest {
@@ -22,12 +24,12 @@ public class ValidationPerformerFactoryTest {
     }
 
     private static void assertGetPerformer(Class expectedPerformerClass, Annotation annotation) {
-        ValidationPerformer performer = ValidationPerformerFactory.getPerformerFor(annotation);
+        IValidationPerformer performer = ValidationPerformerFactory.getPerformerFor(annotation);
         assertEquals(expectedPerformerClass, performer.getClass());
     }
 
-    /** Public method to be used in every ValidationPerformer test. */
-    public static void assertGetPerformer(ValidationPerformer expectedPerformer, Annotation annotation) {
+    /** Public method to be used in every IValidationPerformer test. */
+    public static void assertGetPerformer(IValidationPerformer expectedPerformer, Annotation annotation) {
         assertGetPerformer(expectedPerformer.getClass(), annotation);
     }
 }
