+++ core/src/test/java/net/sourceforge/html5val/util/MockSizeBuilder.java
@@ -1,0 +1,20 @@
+package net.sourceforge.html5val.util;
+
+public class MockSizeBuilder {
+
+    private MockSize size = new MockSize();
+
+    public MockSizeBuilder withMin(int min) {
+        size.setMin(min);
+        return this;
+    }
+
+    public MockSizeBuilder withMax(int max) {
+        size.setMax(max);
+        return this;
+    }
+
+    public MockSize build() {
+        return size;
+    }
+}
